Add arguments to TokenService.requestToken()
Add optional arguments "maxAge" and "expirationThreshold", allowing
user to decide how old a token can be before a new one must be fetched (maxAge) and
how must remaining time a token must have where it is valid (expirationThreshold) in
order to be kept in us.

Updating logic is put inside a try-catch block, handling failures in request for new token,
which enables us to reuse an existing token if present, until a new one can be received.